+++ knightbus-azurestorage/src/KnightBus.Azure.Storage/BlobStorageMessageAttachmentProvider.cs
@@ -1,4 +1,7 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Azure;
@@ -12,6 +15,8 @@ namespace KnightBus.Azure.Storage;
 public class BlobStorageMessageAttachmentProvider : IMessageAttachmentProvider
 {
     private readonly string _connectionString;
+    internal const string FileNameKey = "Filename";
+    private static readonly HashSet<string> Keys = [FileNameKey];
 
     public BlobStorageMessageAttachmentProvider(string connectionString)
     {
@@ -27,7 +32,12 @@ public class BlobStorageMessageAttachmentProvider : IMessageAttachmentProvider
         var blob = new BlobClient(_connectionString, queueName, id);
         var properties = await blob.GetPropertiesAsync(cancellationToken: cancellationToken).ConfigureAwait(false);
 
-        return new MessageAttachment(properties.Value.Metadata["Filename"], properties.Value.ContentType, await blob.OpenReadAsync(cancellationToken: cancellationToken).ConfigureAwait(false));
+        return new MessageAttachment(
+            properties.Value.Metadata[FileNameKey],
+            properties.Value.ContentType,
+            await blob.OpenReadAsync(cancellationToken: cancellationToken).ConfigureAwait(false),
+            properties.Value.Metadata.ToDictionary(x => x.Key, x => Keys.Contains(x.Key) ? x.Value : FromBase64(x.Value))
+        );
     }
 
     public async Task UploadAttachmentAsync(string queueName, string id, IMessageAttachment attachment, CancellationToken cancellationToken = default(CancellationToken))
@@ -35,7 +45,15 @@ public class BlobStorageMessageAttachmentProvider : IMessageAttachmentProvider
         var blob = new BlobClient(_connectionString, queueName, id);
         try
         {
-            await blob.UploadAsync(attachment.Stream, new BlobHttpHeaders { ContentType = attachment.ContentType }, new Dictionary<string, string> { { "Filename", attachment.Filename } }, cancellationToken: cancellationToken)
+            var requiredMetadata = new Dictionary<string, string> { { FileNameKey, attachment.Filename } };
+            var userMetadata = attachment.Metadata.ToDictionary(
+                x => x.Key,
+                x => ToBase64(x.Value));
+
+            var metadata = new Dictionary<string, string>(userMetadata);
+            requiredMetadata.ToList().ForEach(x => metadata[x.Key] = x.Value); // Merge the dictionaries, on collisions, override keys in user's metadata with requiredMetadata
+            
+            await blob.UploadAsync(attachment.Stream, new BlobHttpHeaders { ContentType = attachment.ContentType }, metadata, cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
         }
         catch (RequestFailedException e) when (e.Status == 404)
@@ -67,4 +85,8 @@ public class BlobStorageMessageAttachmentProvider : IMessageAttachmentProvider
             return false;
         }
     }
+    
+    private static string ToBase64(string str) => Convert.ToBase64String(Encoding.UTF8.GetBytes(str));
+    private static string FromBase64(string str) => Encoding.UTF8.GetString(Convert.FromBase64String(str));
+
 }