+++ knightbus/src/KnightBus.Core/MessageAttachment.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using KnightBus.Messages;
 
@@ -6,11 +7,12 @@ namespace KnightBus.Core;
 
 public class MessageAttachment : IMessageAttachment
 {
-    public MessageAttachment(string filename, string contentType, Stream stream)
+    public MessageAttachment(string filename, string contentType, Stream stream, Dictionary<string,string> metadata = null)
     {
         Filename = filename;
         ContentType = contentType;
         Stream = stream;
+        Metadata = metadata ?? [];
         try
         {
             Length = stream?.Length ?? 0;
@@ -23,4 +25,5 @@ public class MessageAttachment : IMessageAttachment
     public string ContentType { get; protected set; }
     public long Length { get; protected set; }
     public Stream Stream { get; protected set; }
+    public Dictionary<string, string> Metadata { get; protected set; }
 }